t/fast-testing.sh [--coproc] cproc optional
coproc only works with bash-4, and it kills all processes with matching names
even in other ttys

+++ t/fast-testing.sh
@@ -1,4 +1,6 @@
 #!/bin/bash
+# usage: t/fast-testing.sh [--coproc]
+
 test -f Makefile || perl Makefile.PL
 V=`perl -ane'print $F[2] if /^VERSION =/' Makefile`
 R=`git log -1 --pretty=format:"%h"`
@@ -8,18 +10,22 @@ trap "rm $lock; exit 255" SIGINT SIGTERM
 
 # need to kill rogue processes since we cannot use run_cmd. it disturbs stdout/err order #
 # bash-4 only
-coproc (while true; do
+if [ x$1 = x--coproc ]; then
+  coproc (while true; do
     sleep 1;
     code=`ps axw|egrep ' \./(ccode|cccode|a |aa |a.out|perldoc)'|grep -v grep`
     pid=`echo $code|perl -ane'print $F[0]'`
     test -n "$pid" && (echo $code; sleep 1s; kill $pid 2>/dev/null);
     sleep 5; done)
-w=${COPROC_PID}
+  w=${COPROC_PID}
+fi
 
 # test locally, ~5:45hr (17*20min)
-PERLCC_TIMEOUT=15 NO_AUTHOR=1 perlall -m make '-S prove -b -j4'
+PERLCC_TIMEOUT=15 NO_AUTHOR=1 perlall -mq make '-S prove -b -j4'
 
-kill -9 $w
+if [ x$1 = x--coproc ]; then
+  kill -9 $w
+fi
 
 # creates log.modules files with date added
 # perlall -m make '-Iblib/arch -Iblib/lib t/modules.t -no-subset -no-date t/top100'
@@ -35,5 +41,4 @@ if [ -n "$logs" ]; then
     cp status.$V-$R $rdir/
 fi
 rm $lock
-
